Move settings to config
